PDFIssue metadata extraction methods refactored into mixin.
